requirements.txt
python-telegram-bot==21.7
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
aiohttp==3.9.1
python-dotenv==1.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
pytz==2023.3
geopy==2.4.0
tzwhere==3.0.3
timezonefinderl
dateparser